Group manage GDPR update
